Add automatic Release mechanism
Now when COM objects are returned to Python-land they will automatically
have Release() called when Python garbage collects them. One exception
is IBlackmagicRawJob, which is not managed automatically since there
seems to be a requirement that jobs are released before submitted.
Unfortunately this prevents us from using the automatic mechanism since
we can't guarantee that the GC will run (and release the job) when it is
manually deleted using `del`.